## Current Issue
Mitigation to avoid blocking copy operations across different CUDA streams. Details could be found in: https://github.com/vllm-project/vllm/issues/22754

## Change
When we copy the sampled valid token ids from device to host, avoid using `tolist` which would trigger a CUDA wise stream sync if the source is on device. We change it to use non-blocking copy followed by an explicit CUDA event sync.

## Test for Non Disagg
Bring up vLLM server
```
VLLM_USE_V1=1 vllm serve Qwen/Qwen2.5-14B-Instruct --disable-l
og-requests -tp 8 --max-num-seqs 64 --no-enable-prefix-caching --max_num_batched_tokens=8000
```

Load test for non-disagg setup and observe no perf difference:

Before
```
Maximum request concurrency: 64
============ Serving Benchmark Result ============
Successful requests:                     640       
Maximum request concurrency:             64        
Benchmark duration (s):                  29.60     
Total input tokens:                      1276994   
Total generated tokens:                  93700     
Request throughput (req/s):              21.62     
Output token throughput (tok/s):         3165.99   
Total Token throughput (tok/s):          46313.80  
---------------Time to First Token----------------
Mean TTFT (ms):                          327.88    
Median TTFT (ms):                        302.90    
P99 TTFT (ms):                           1566.85   
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          18.16     
Median TPOT (ms):                        18.01     
P99 TPOT (ms):                           20.10     
---------------Inter-token Latency----------------
Mean ITL (ms):                           17.82     
Median ITL (ms):                         9.49      
P99 ITL (ms):                            102.45    
==================================================
``` 

After
```
Maximum request concurrency: 64
============ Serving Benchmark Result ============
Successful requests:                     640       
Maximum request concurrency:             64        
Benchmark duration (s):                  29.58     
Total input tokens:                      1276994   
Total generated tokens:                  93693     
Request throughput (req/s):              21.64     
Output token throughput (tok/s):         3167.38   
Total Token throughput (tok/s):          46337.38  
---------------Time to First Token----------------
Mean TTFT (ms):                          320.65    
Median TTFT (ms):                        302.65    
P99 TTFT (ms):                           1568.02   
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          18.21     
Median TPOT (ms):                        18.17     
P99 TPOT (ms):                           20.09     
---------------Inter-token Latency----------------
Mean ITL (ms):                           17.86     
Median ITL (ms):                         9.50      
P99 ITL (ms):                            102.48    
==================================================
```


## Test for Disagg
Runs with a vendor internal disagg implementation (kv connector interface based), and observe TTIT wins.

Before
```
Ran 1029/1029 requests in 133.00s
Success rate:        100.00%
QPS:                 7.74
Avg latency:         4.841s
Avg TTFT (client):   137.37ms
P50 TTFT (client):   130.12ms
P99 TTFT (client):   364.42ms
Avg TTIT (client):   31.35ms
P50 TTIT (client):   31.12ms
P99 TTIT (client):   38.27ms
Avg TTFT (server):   134.39ms
Avg TTIT (server):   31.29ms
Avg prefill len:     2199.72 tokens
P50 prefill len:     2199.00 tokens
P99 prefill len:     2231.00 tokens
Avg decode len:      150.00 tokens
P50 decode len:      150.00 tokens
P99 decode len:      150.00 tokens
```
After
```
Ran 1026/1026 requests in 132.34s
Success rate:        100.00%
QPS:                 7.75
Avg latency:         4.550s
Avg TTFT (client):   128.50ms
P50 TTFT (client):   125.91ms
P99 TTFT (client):   205.68ms
Avg TTIT (client):   29.47ms
P50 TTIT (client):   29.59ms
P99 TTIT (client):   30.00ms
Avg TTFT (server):   124.88ms
Avg TTIT (server):   28.88ms
Avg prefill len:     2199.10 tokens
P50 prefill len:     2199.00 tokens
P99 prefill len:     2233.00 tokens
Avg decode len:      150.00 tokens
P50 decode len:      150.00 tokens
P99 decode len:      150.00 tokens
```

Also confirmed from GPU trace that there is no blocking behavior between vLLM model forward CUDA stream with other CUDA streams.
<img width="1211" height="1926" alt="Screenshot 2025-08-12 at 11 28 46 AM" src="https://github.com/user-attachments/assets/b8940f0a-75c2-4660-a0d9-966622437b5b" />


## Accuracy
gsm8K 8shot (disagg)
```
[2025-07-30 23:06:36,254] [rank 0] [INFO] Per prompt detailed info dumped to /tmp/eval_dump.gsm8k.8_shot.1_gen.20250730_230636.json
[2025-07-30 23:06:36,254] [rank 0] [INFO] Evaluation results on task gsm8k.8_shot.1_gen: em: 0.970000 | f1: 0.970000 | em_maj1@1: 0.970000 | f1_maj1@1: 0.970000
[2025-07-30 23:06:36,254] [rank 0] [INFO] Task gsm8k.8_shot.1_gen took 54.20 seconds
```

mmlu_pro (non-disagg)
```
vllm (pretrained=/data/users/zijingliu/cp/Llama-4-Maverick-17B-128E-Instruct-FP8,dtype=auto,max_model_len=8196,tensor_parallel_size=8,gpu_memory_utilization=0.9,max_gen_toks=2048,seed=0), gen_kwargs: (None), limit: 500.0, num_fewshot: None, batch_size: auto
|       Tasks       |Version|    Filter    |n-shot|  Metric   |   |Value |   |Stderr|
|-------------------|------:|--------------|-----:|-----------|---|-----:|---|-----:|
|mmlu_pro           |    2.0|custom-extract|      |exact_match|↑  |0.7994|±  |0.0048|
| - biology         |    2.1|custom-extract|     5|exact_match|↑  |0.8960|±  |0.0137|
| - business        |    2.1|custom-extract|     5|exact_match|↑  |0.8460|±  |0.0162|
| - chemistry       |    2.1|custom-extract|     5|exact_match|↑  |0.8540|±  |0.0158|
| - computer_science|    2.1|custom-extract|     5|exact_match|↑  |0.8390|±  |0.0182|
| - economics       |    2.1|custom-extract|     5|exact_match|↑  |0.8520|±  |0.0159|
| - engineering     |    2.1|custom-extract|     5|exact_match|↑  |0.7300|±  |0.0199|
| - health          |    2.1|custom-extract|     5|exact_match|↑  |0.7640|±  |0.0190|
| - history         |    2.1|custom-extract|     5|exact_match|↑  |0.7060|±  |0.0234|
| - law             |    2.1|custom-extract|     5|exact_match|↑  |0.5960|±  |0.0220|
| - math            |    2.1|custom-extract|     5|exact_match|↑  |0.9020|±  |0.0133|
| - other           |    2.1|custom-extract|     5|exact_match|↑  |0.7600|±  |0.0191|
| - philosophy      |    2.1|custom-extract|     5|exact_match|↑  |0.7375|±  |0.0197|
| - physics         |    2.1|custom-extract|     5|exact_match|↑  |0.8680|±  |0.0152|
| - psychology      |    2.1|custom-extract|     5|exact_match|↑  |0.8260|±  |0.0170|

| Groups |Version|    Filter    |n-shot|  Metric   |   |Value |   |Stderr|
|--------|------:|--------------|------|-----------|---|-----:|---|-----:|
|mmlu_pro|      2|custom-extract|      |exact_match|↑  |0.7994|±  |0.0048|
```



cc @WoosukKwon  @lucas-tucker  @simon-mo @houseroad