@njhill 2025-08-29
> @liuzijing2014 @njhill Is this the way to transfer tensors from GPU to CPU in asynchronously in vLLM? I'm curious for the spec decoding code path. I'm a bit confused about the fact that there is still a `.tolist()` call. How come this doesn't block, but it did before? And what about the event object? Perhaps each `tolist()` call should create its own CUDA event object?

@tomasruizt it still blocks, it's just that it can now happen concurrently with transfers that kv connectors might make in different cuda streams.

We are working on eliminating this sync point altogether, including for the spec decoding case: https://github.com/vllm-project/vllm/pull/23569